requirements.txt
Django==3.1.7
djangorestframework==3.12.4
djoser==2.1.0
djangorestframework-simplejwt==4.6.0
